backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - Simplified for MVP (SQLite)
sqlalchemy==2.0.25

# Telegram Bot
python-telegram-bot==20.7

# HTTP Client
httpx==0.25.2

# Timezone handling
pytz==2024.1

# Utilities
python-dotenv==1.0.1
